refactor: split src/context.rs into src/context/ module tree
Extract budgeting and rendering helpers from the monolithic src/context.rs
(~8400 lines) into focused submodules as the first phase of the module split
recommended in #1892.

- src/context/budget.rs: token estimation and budget-fitting helpers
  (estimate_text_tokens, estimate_section_tokens, estimate_message_tokens,
  truncate_section_content, fit_section_to_budget)

- src/context/render.rs: generic message/body/header rendering and label
  helpers (message_header, render_message, body_preview, sanitize_inline,
  indent_block, enum_label, all *_label functions, etc.)

- src/context/mod.rs: retains the main build_context pipeline, work-item
  projection, recent-turn projection, episode memory, and all existing tests

Adds a regression test verifying the moved budget helpers produce consistent
output after the split.

Refs: #1892
